@@ -3,7 +3,7 @@ publish_to: none
 
 environment:
   sdk: ^3.0.0
-  flutter: '>=3.10.0'
+  flutter: '>=3.27.0-0'
 
 dependencies:
   align_positioned: '>=3.0.0 <5.0.0'
